Adding staticL1Test
